requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
dlib==20.0.0
facenet-pytorch==2.6.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.1
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.5
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==10.2.0
protobuf==4.25.8
Pygments==2.19.2
requests==2.32.4
rich==14.1.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
